move from frontend/osoc-6 to frontend
