+++ frontend/next.config.js
