+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "osoc-6",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "scripts": {
